+++ static/app/views/onboarding/targetedOnboarding/platform.tsx
@@ -10,7 +10,8 @@ import testableTransition from 'sentry/utils/testableTransition';
 import StepHeading from 'sentry/views/onboarding/components/stepHeading';
 
 import CreateProjectsFooter from './components/createProjectsFooter';
-import {StepProps, usePersistedOnboardingState} from './types';
+import {StepProps} from './types';
+import {usePersistedOnboardingState} from './utils';
 
 function OnboardingPlatform(props: StepProps) {
   const [selectedPlatforms, setSelectedPlatforms] = useState<PlatformKey[]>([]);
